implement CLI interface with CommandRunner (using sample in https://pub.dev/packages/args#dispatching-commands ) also exclude [packageLocation]/lib/src/*.dart when searching for libraries
