+++ lib/src/package_analysis/common.dart
@@ -1,0 +1,6 @@
+import 'dart:io';
+
+void exitFatal(String errorMessage, int exitCode) {
+  stderr.writeln(errorMessage);
+  exit(exitCode);
+}
